+++ braintree/paypal_here.py
@@ -1,0 +1,7 @@
+import braintree
+from braintree.resource import Resource
+
+class PayPalHere(Resource):
+    def __init__(self, gateway, attributes):
+        Resource.__init__(self, gateway, attributes)
+