+++ tests/unit/test_credit_card_verification.py
@@ -34,6 +34,21 @@ class TestCreditCardVerification(unittest.TestCase):
         verification = CreditCardVerification(None, {"amount": "1.00"})
         self.assertEqual(verification.risk_data, None)
 
+    def test_constructor_when_network_response_is_included(self):
+        attributes = {
+            'amount': '1.00',
+            'network_response_code': '00',
+            'network_response_text': 'Successful approval/completion or V.I.P. PIN verification is successful'
+        }
+        verification = CreditCardVerification(None, attributes)
+        self.assertEqual(verification.network_response_code, '00')
+        self.assertEqual(verification.network_response_text, 'Successful approval/completion or V.I.P. PIN verification is successful')
+
+    def test_constructor_when_network_response_is_not_included(self):
+        verification = CreditCardVerification(None, {'amount': '1.00'})
+        self.assertEqual(verification.network_response_code, None)
+        self.assertEqual(verification.network_response_text, None)
+
     @raises(NotFoundError)
     def test_finding_empty_id_raises_not_found_exception(self):
         CreditCardVerification.find(" ")